feat: add "add items" button

+++ Web.BlazorServer/ViewModels/Transaction/StockTransferRequest/StockTransferRequestLineVM.cs
@@ -6,7 +6,7 @@ public class StockTransferRequestLineVM
     public string ItemDescription { get; set; } = string.Empty;
     public string UoM { get; set; } = string.Empty;
     public string Warehouse { get; set; } = string.Empty;
-    public int QuantityOnHand { get; set; }
-    public int QuantityAlloted { get; set; }
+    public decimal QuantityOnHand { get; set; }
+    public decimal QuantityAlloted { get; set; }
 
 }
